feat : 주문 처리중 실패하는 경우(재고 부족, 쿠폰 사용 처리 실패 등등..) 주문 자체를 실패처리 하기 위한 별도의 서비스 생성

+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/payment/PaymentGatewayRestPort.java
@@ -27,7 +27,7 @@ public class PaymentGatewayRestPort implements PaymentGatewayPort {
     public PgPaymentResponse requestPayment(String orderKey,
                                             CardType cardType,
                                             String cardNo,
-                                            BigDecimal amount,
+                                            Long amount,
                                             Long memberId) {
 
         HttpHeaders headers = new HttpHeaders();
